+++ DataAccess/Repository/IRepository/ICategoryRepository.cs
@@ -10,8 +10,8 @@ namespace DataAccess.Repository.IRepository
     public interface ICategoryRepository
     {
         Task<List<Category>> GetAll();
-        Task<bool> CheckProductExistsByCateId(int categoryId);
-        Task<Category?> FindById(int categoryId);
+        Task<bool> CheckProductExistsByCateId(long categoryId);
+        Task<Category?> FindById(long categoryId);
         Task<Category?> FindByName(string categoryName);
         Task Add(Category category);
         Task Update(Category category);
