+++ DataAccess/Repository/NotificationRepository.cs
@@ -6,7 +6,7 @@ namespace DataAccess.Repository
 {
     public class NotificationRepository : INotificationRepository
     {
-        public async Task<List<Notification>> GetAll(int userId)
+        public async Task<List<Notification>> GetAll(long userId)
             => await NotificationDAO.GetAll(userId);
 
         public async Task Update(Notification notification)
